+++ example/integration_test/scenario_device_id_init/10.change_with_merge_test.dart
@@ -3,18 +3,22 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import '../utils.dart';
 
-/// 5.Check if setDeviceId() sets the device ID correctly and the type is DEVELOPER_SUPPLIED using deviceID module
+/// 10.Testing changeWithMerge forces merge on the server.
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
-  testWidgets('5.Test Device ID Type - Developer sets the ID during init using Device ID Module', (WidgetTester tester) async {
+  testWidgets('10.Testing changeWithMerge forces merge on the server.', (WidgetTester tester) async {
     // Initialize the SDK
     CountlyConfig config = CountlyConfig(SERVER_URL, APP_KEY).setLoggingEnabled(true).setDeviceId('test');
     await Countly.initWithConfig(config);
+    await Countly.instance.deviceId.changeWithMerge('test1');
     // Get the device ID type
     DeviceIdType? type = await Countly.instance.deviceId.getIDType();
     String? id = await Countly.instance.deviceId.getID();
     // Verify the device ID type
     expect(type, DeviceIdType.DEVELOPER_SUPPLIED);
-    expect(id, 'test');
+    expect(id, 'test1');
+
+    // Check that deviceID parameter is set correctly in request queue
+    await testCustomRequestParams({'old_device_id': 'test', 'device_id': 'test1'});
   });
 }