+++ lib/src/countly_config.dart
@@ -145,9 +145,7 @@ class CountlyConfig {
 
   /// unique ID for the device the app is running on.
   CountlyConfig setDeviceId(String deviceID) {
-    if (deviceID.isNotEmpty) {
-      _deviceID = deviceID;
-    }
+    _deviceID = deviceID;
     return this;
   }